Simplified data source structure and notifications.
