+++ docs/background/faq.rst
