+++ src/core/tests/sources/test_tree_source.py
@@ -1,13 +1,123 @@
 from unittest import TestCase
 from unittest.mock import Mock
 
-from toga.sources import DictSource
+from toga.sources import TreeSource
+from toga.sources.tree_source import Node
 
 
-class DictSourceTests(TestCase):
+class LeafNodeTests(TestCase):
+    def setUp(self):
+        self.source = Mock()
+        self.example = Node(val1='value 1', val2=42)
+        self.example._source = self.source
+
+    def test_initial_state(self):
+        "A node holds values as expected"
+        self.assertEqual(self.example.val1, 'value 1')
+        self.assertEqual(self.example.val2, 42)
+        self.assertFalse(self.example.has_children())
+        self.assertEqual(len(self.example), 0)
+
+    def test_change_value(self):
+        "If a node value changes, the source is notified"
+        self.example.val1 = 'new value'
+
+        self.assertEqual(self.example.val1, 'new value')
+        self.source._notify.assert_called_once_with('change', item=self.example)
+
+
+class NodeTests(TestCase):
+    def setUp(self):
+        self.source = Mock()
+
+        def bound_create_node(s):
+            def create_node(value):
+                return Node(source=s, **value)
+            return create_node
+
+        self.source._create_node = bound_create_node(self.source)
+
+        self.parent = Node(val1='value 1', val2=42)
+        self.parent._source = self.source
+        self.parent._children = []
+        for datum in [{'val1': 'child 1', 'val2': 11}, {'val1': 'child 2', 'val2': 22}]:
+            child = Node(**datum)
+            child.source = self.source
+            self.parent._children.append(child)
+
+    def test_initial_state(self):
+        "A node holds values as expected"
+
+        self.assertEqual(self.parent.val1, 'value 1')
+        self.assertEqual(self.parent.val2, 42)
+        self.assertTrue(self.parent.has_children())
+        self.assertEqual(len(self.parent), 2)
+
+    def test_change_value(self):
+        "If a node value changes, the source is notified"
+        self.parent.val1 = 'new value'
+
+        self.assertEqual(self.parent.val1, 'new value')
+        self.source._notify.assert_called_once_with('change', item=self.parent)
+
+    def test_empty_children(self):
+        "A parent with 0 children isn't the same as a parent who *can't* have children"
+        parent = Node(source=self.source, val1='value 1', val2=42)
+        parent._children = []
+
+        self.assertTrue(parent.has_children())
+        self.assertEqual(len(parent), 0)
+
+    def test_change_child(self):
+        "Changing a child notifies the source"
+        # Check initial value
+        self.assertEqual(len(self.parent), 2)
+        self.assertEqual(self.parent[1].val1, 'child 2')
+        self.assertEqual(self.parent[1].val2, 22)
+
+        # Change the value
+        self.parent[1] = {'val1': 'new child', 'val2': 33}
+
+        # Check the values after modification
+        self.assertEqual(len(self.parent), 2)
+        self.assertEqual(self.parent[1].val1, 'new child')
+        self.assertEqual(self.parent[1].val2, 33)
+
+    def test_insert_child(self):
+        "A new child can be inserted; defers to the source"
+        self.parent.insert(1, val1='inserted 1', val2=33)
+        self.source.insert.assert_called_once_with(self.parent, 1, val1='inserted 1', val2=33)
+
+    def test_append_child(self):
+        "A new child can be appended; defers to the source"
+        self.parent.append(val1='appended 1', val2=33)
+        self.source.append.assert_called_once_with(self.parent, val1='appended 1', val2=33)
+
+    def test_prepend_child(self):
+        "A new child can be prepended; defers to the source"
+        self.parent.prepend(val1='prepended 1', val2=33)
+        self.source.prepend.assert_called_once_with(self.parent, val1='prepended 1', val2=33)
+
+    def test_remove_child(self):
+        "A child can be removed; defers to the source"
+        child = self.parent[1]
+        self.parent.remove(child)
+        self.source.remove.assert_called_once_with(self.parent, child)
+
+    def test_iterate_children(self):
+        "Children of a node can be iterated over"
+        result = 0
+
+        for child in self.parent:
+            result += child.val2
+
+        self.assertEqual(result, 33)
+
+
+class TreeSourceTests(TestCase):
     def test_init_with_list_of_tuples(self):
-        "DictSources can be instantiated from lists of tuples"
-        source = DictSource(
+        "TreeSources can be instantiated from lists of tuples"
+        source = TreeSource(
             data=[
                 ('first', 111),
                 ('second', 222),
@@ -44,8 +154,8 @@ class DictSourceTests(TestCase):
         listener.insert.assert_called_once_with(parent=None, index=1, item=source[1])
 
     def test_init_with_list_of_dicts(self):
-        "DictSource nodes can be instantiated from lists of dicts"
-        source = DictSource(
+        "TreeSource nodes can be instantiated from lists of dicts"
+        source = TreeSource(
             data=[
                 {'val1': 'first', 'val2': 111},
                 {'val1': 'second', 'val2': 222},
@@ -82,8 +192,8 @@ class DictSourceTests(TestCase):
         listener.insert.assert_called_once_with(parent=None, index=1, item=source[1])
 
     def test_init_with_dict_of_lists(self):
-        "DictSource nodes can be instantiated from dicts of lists"
-        source = DictSource(
+        "TreeSource nodes can be instantiated from dicts of lists"
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -141,8 +251,8 @@ class DictSourceTests(TestCase):
         listener.insert.assert_called_once_with(parent=None, index=2, item=source[2])
 
     def test_init_with_dict_of_dicts(self):
-        "DictSource nodes can be instantiated from dicts of dicts"
-        source = DictSource(
+        "TreeSource nodes can be instantiated from dicts of dicts"
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -208,7 +318,7 @@ class DictSourceTests(TestCase):
 
     def test_iter(self):
         "TreeSource roots can be iterated over"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -228,7 +338,7 @@ class DictSourceTests(TestCase):
 
     def test_insert_root_args(self):
         "A new root can be inserted using value args"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -257,7 +367,7 @@ class DictSourceTests(TestCase):
 
     def test_insert_root_kwargs(self):
         "A new root can be inserted using kwargs"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -286,7 +396,7 @@ class DictSourceTests(TestCase):
 
     def test_insert_child_args(self):
         "A new child can be inserted using value args"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -317,7 +427,7 @@ class DictSourceTests(TestCase):
 
     def test_insert_child_kwargs(self):
         "A new child can be inserted using kwargs"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -348,7 +458,7 @@ class DictSourceTests(TestCase):
 
     def test_insert_first_child(self):
         "If a node previously didn't allow children, inserting changes this"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -381,7 +491,7 @@ class DictSourceTests(TestCase):
 
     def test_append_root(self):
         "A new root can be appended"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -410,7 +520,7 @@ class DictSourceTests(TestCase):
 
     def test_append_child(self):
         "A new child can be appended"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -441,7 +551,7 @@ class DictSourceTests(TestCase):
 
     def test_prepend_root(self):
         "A new root can be prepended"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -470,7 +580,7 @@ class DictSourceTests(TestCase):
 
     def test_prepend_child(self):
         "A new child can be prepended"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -501,7 +611,7 @@ class DictSourceTests(TestCase):
 
     def test_remove_root(self):
         "A root can be removed"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
@@ -529,7 +639,7 @@ class DictSourceTests(TestCase):
 
     def test_remove_child(self):
         "A child can be removed"
-        source = DictSource(
+        source = TreeSource(
             data={
                 ('first', 111): None,
                 ('second', 222): [],
