+++ src/core/tests/widgets/test_detailedlist.py
