+++ src/dummy/toga_dummy/widgets/table.py
@@ -8,14 +8,14 @@ class Table(Widget):
     def change_source(self, source):
         self._action('change source', source=source)
 
-    def insert(self, index, row):
-        self._action('insert row', index=index, row=row)
+    def insert(self, item):
+        self._action('insert row', item=item)
 
-    def change(self, row):
-        self._action('change row', row=row)
+    def change(self, item):
+        self._action('change row', item=item)
 
-    def remove(self, row):
-        self._action('remove row', row=row)
+    def remove(self, item):
+        self._action('remove row', item=item)
 
     def clear(self):
         self._action('clear')
