+++ src/dummy/toga_dummy/widgets/tree.py
@@ -8,14 +8,14 @@ class Tree(Widget):
     def change_source(self, source):
         self._action('change source', source=source)
 
-    def insert(self, parent, index, node):
-        self._action('insert node', parent=parent, index=index, node=node)
+    def insert(self, item):
+        self._action('insert node', item=item)
 
-    def change(self, node):
-        self._action('change node', node=node)
+    def change(self, item):
+        self._action('change node', item=item)
 
-    def remove(self, node):
-        self._action('remove node', node=node)
+    def remove(self, item):
+        self._action('remove node', item=item)
 
     def clear(self):
         self._action('clear')
